Move token controller to Typescript
